python-server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
# 実際の実装では以下も必要
# openai==0.28.1
